# Slightly optimize very short queries with LowCardinality