@CurtizJ 2020-08-26
Look at QPS.
Before:
```
echo "SELECT col1, col2 FROM test PREWHERE col1 = '042fecec-ea68-1b26-00a4-a4f3b3987574'" | clickhouse-benchmark -c 1 -i 2000 -d 0 
Loaded 1 queries.

Queries executed: 2000.

localhost:9000, queries 2000, QPS: 735.422, RPS: 94134.037, MiB/s: 1.436, result RPS: 0.000, result MiB/s: 0.000.

0.000%		0.001 sec.	
10.000%		0.001 sec.	
20.000%		0.001 sec.	
30.000%		0.001 sec.	
40.000%		0.001 sec.	
50.000%		0.001 sec.	
60.000%		0.001 sec.	
70.000%		0.001 sec.	
80.000%		0.002 sec.	
90.000%		0.002 sec.	
95.000%		0.002 sec.	
99.000%		0.002 sec.	
99.900%		0.004 sec.	
99.990%		0.011 sec.	
```

After:
```
echo "SELECT col1, col2 FROM test PREWHERE col1 = '042fecec-ea68-1b26-00a4-a4f3b3987574'" | clickhouse-benchmark -c 1 -i 2000 -d 0 
Loaded 1 queries.

Queries executed: 2000.

localhost:9000, queries 2000, QPS: 1030.173, RPS: 131862.092, MiB/s: 2.012, result RPS: 0.000, result MiB/s: 0.000.

0.000%		0.001 sec.	
10.000%		0.001 sec.	
20.000%		0.001 sec.	
30.000%		0.001 sec.	
40.000%		0.001 sec.	
50.000%		0.001 sec.	
60.000%		0.001 sec.	
70.000%		0.001 sec.	
80.000%		0.001 sec.	
90.000%		0.001 sec.	
95.000%		0.001 sec.	
99.000%		0.002 sec.	
99.900%		0.003 sec.	
99.990%		0.009 sec.
```