@CurtizJ 2020-08-26
Currently it's impossible to add performance test for such short queries.