@alexey-milovidov 2020-08-26
I've read the fix but I don't understand how it's related to LowCardinality.