@CurtizJ 2020-08-26
> I've read the fix but I don't understand how it's related to LowCardinality

The problem appeared, when we read dictionary of low cardinality. Since it has low size, a lot of marks point to one place, and previously we a lot of time was spent in loop, while trying to find mark to next compressed block.