@CurtizJ 2020-08-26
> After your change, is there still a difference with experimental_use_processors?

I can't check it now, because in new versions old code is already removed and `experimental_use_processors` is exists just for backward compatibility and does nothing.